+++ arcade/gl/framebuffer.py
@@ -1,22 +1,19 @@
 from __future__ import annotations
 
-import weakref
-from contextlib import contextmanager
-from ctypes import Array, c_int, c_uint, string_at
-from typing import TYPE_CHECKING, Generator
+from abc import ABC, abstractmethod
 
-from pyglet import gl
+from contextlib import contextmanager
+from typing import TYPE_CHECKING, Generator
 
 from arcade.types import RGBOrA255, RGBOrANormalized
 
 from .texture import Texture2D
-from .types import pixel_formats
 
 if TYPE_CHECKING:
     from arcade.gl import Context
 
 
-class Framebuffer:
+class Framebuffer(ABC):
     """
     An offscreen render target also called a Framebuffer Object in OpenGL.
     This implementation is using texture attachments. When creating a
@@ -51,7 +48,6 @@ class Framebuffer:
     is_default = False
     __slots__ = (
         "_ctx",
-        "_glo",
         "_width",
         "_height",
         "_color_attachments",
@@ -72,7 +68,6 @@ class Framebuffer:
         color_attachments: Texture2D | list[Texture2D],
         depth_attachment: Texture2D | None = None,
     ):
-        self._glo = fbo_id = gl.GLuint()  # The OpenGL alias/name
         self._ctx = ctx
         if not color_attachments:
             raise ValueError("Framebuffer must at least have one color attachment")
@@ -85,10 +80,6 @@ class Framebuffer:
         self._depth_mask = True  # Determines if the depth buffer should be affected
         self._prev_fbo = None
 
-        # Create the framebuffer object
-        gl.glGenFramebuffers(1, self._glo)
-        gl.glBindFramebuffer(gl.GL_FRAMEBUFFER, self._glo)
-
         # Ensure all attachments have the same size.
         # OpenGL do actually support different sizes,
         # but let's keep this simple with high compatibility.
@@ -96,55 +87,10 @@ class Framebuffer:
         self._viewport = 0, 0, self._width, self._height
         self._scissor: tuple[int, int, int, int] | None = None
 
-        # Attach textures to it
-        for i, tex in enumerate(self._color_attachments):
-            # TODO: Possibly support attaching a specific mipmap level
-            #       but we can read from specific mip levels from shaders.
-            gl.glFramebufferTexture2D(
-                gl.GL_FRAMEBUFFER,
-                gl.GL_COLOR_ATTACHMENT0 + i,
-                tex._target,
-                tex.glo,
-                0,  # Level 0
-            )
-
-        if self.depth_attachment:
-            gl.glFramebufferTexture2D(
-                gl.GL_FRAMEBUFFER,
-                gl.GL_DEPTH_ATTACHMENT,
-                self.depth_attachment._target,
-                self.depth_attachment.glo,
-                0,
-            )
-
-        # Ensure the framebuffer is sane!
-        self._check_completeness()
-
-        # Set up draw buffers. This is simply a prepared list of attachments enums
-        # we use in the use() method to activate the different color attachment layers
-        layers = [gl.GL_COLOR_ATTACHMENT0 + i for i, _ in enumerate(self._color_attachments)]
-        # pyglet wants this as a ctypes thingy, so let's prepare it
-        self._draw_buffers: Array[c_uint] | None = (gl.GLuint * len(layers))(*layers)
-
-        # Restore the original bound framebuffer to avoid confusion
-        self.ctx.active_framebuffer.use(force=True)
-
-        if self._ctx.gc_mode == "auto" and not self.is_default:
-            weakref.finalize(self, Framebuffer.delete_glo, ctx, fbo_id)
-
         self.ctx.stats.incr("framebuffer")
 
-    def __del__(self):
-        # Intercept garbage collection if we are using Context.gc()
-        if self._ctx.gc_mode == "context_gc" and not self.is_default and self._glo.value > 0:
-            self._ctx.objects.append(self)
-
     @property
-    def glo(self) -> gl.GLuint:
-        """The OpenGL id/name of the framebuffer."""
-        return self._glo
-
-    def _get_viewport(self) -> tuple[int, int, int, int]:
+    def viewport(self) -> tuple[int, int, int, int]:
         """
         Get or set the framebuffer's viewport.
 
@@ -162,24 +108,13 @@ class Framebuffer:
         """
         return self._viewport
 
-    def _set_viewport(self, value: tuple[int, int, int, int]):
-        if not isinstance(value, tuple) or len(value) != 4:
-            raise ValueError("viewport should be a 4-component tuple")
+    @viewport.setter
+    @abstractmethod
+    def viewport(self, value: tuple[int, int, int, int]):
+        raise NotImplementedError("The enabled graphics backend does not support this method.")
 
-        self._viewport = value
-
-        # If the framebuffer is bound we need to set the viewport.
-        # Otherwise it will be set on use()
-        if self._ctx.active_framebuffer == self:
-            gl.glViewport(*self._viewport)
-            if self._scissor is None:
-                gl.glScissor(*self._viewport)
-            else:
-                gl.glScissor(*self._scissor)
-
-    viewport = property(_get_viewport, _set_viewport)
-
-    def _get_scissor(self) -> tuple[int, int, int, int] | None:
+    @property
+    def scissor(self) -> tuple[int, int, int, int] | None:
         """
         Get or set the scissor box for this framebuffer.
 
@@ -199,17 +134,10 @@ class Framebuffer:
         """
         return self._scissor
 
-    def _set_scissor(self, value):
-        self._scissor = value
-
-        if self._scissor is None:
-            if self._ctx.active_framebuffer == self:
-                gl.glScissor(*self._viewport)
-        else:
-            if self._ctx.active_framebuffer == self:
-                gl.glScissor(*self._scissor)
-
-    scissor = property(_get_scissor, _set_scissor)
+    @scissor.setter
+    @abstractmethod
+    def scissor(self, value):
+        raise NotImplementedError("The enabled graphics backend does not support this method.")
 
     @property
     def ctx(self) -> "Context":
@@ -260,11 +188,9 @@ class Framebuffer:
         return self._depth_mask
 
     @depth_mask.setter
+    @abstractmethod
     def depth_mask(self, value: bool):
-        self._depth_mask = value
-        # Set state if framebuffer is active
-        if self._ctx.active_framebuffer == self:
-            gl.glDepthMask(self._depth_mask)
+        raise NotImplementedError("The enabled graphics backend does not support this method.")
 
     def __enter__(self):
         self._prev_fbo = self._ctx.active_framebuffer
@@ -300,25 +226,12 @@ class Framebuffer:
         self._use(force=force)
         self._ctx.active_framebuffer = self
 
+    @abstractmethod
     def _use(self, *, force: bool = False):
         """Internal use that do not change the global active framebuffer"""
-        if self.ctx.active_framebuffer == self and not force:
-            return
+        raise NotImplementedError("The enabled graphics backend does not support this method.")
 
-        gl.glBindFramebuffer(gl.GL_FRAMEBUFFER, self._glo)
-
-        # NOTE: gl.glDrawBuffer(GL_NONE) if no texture attachments (future)
-        # NOTE: Default framebuffer currently has this set to None
-        if self._draw_buffers:
-            gl.glDrawBuffers(len(self._draw_buffers), self._draw_buffers)
-
-        gl.glDepthMask(self._depth_mask)
-        gl.glViewport(*self._viewport)
-        if self._scissor is not None:
-            gl.glScissor(*self._scissor)
-        else:
-            gl.glScissor(*self._viewport)
-
+    @abstractmethod
     def clear(
         self,
         *,
@@ -350,44 +263,9 @@ class Framebuffer:
             viewport:
                 The viewport range to clear
         """
-        with self.activate():
-            scissor_values = self._scissor
+        raise NotImplementedError("The enabled graphics backend does not support this method.")
 
-            if viewport:
-                self.scissor = viewport
-            else:
-                self.scissor = None
-
-            clear_color = 0.0, 0.0, 0.0, 0.0
-            if color is not None:
-                if len(color) == 3:
-                    clear_color = color[0] / 255, color[1] / 255, color[2] / 255, 1.0
-                elif len(color) == 4:
-                    clear_color = color[0] / 255, color[1] / 255, color[2] / 255, color[3] / 255
-                else:
-                    raise ValueError("Color should be a 3 or 4 component tuple")
-            elif color_normalized is not None:
-                if len(color_normalized) == 3:
-                    clear_color = color_normalized[0], color_normalized[1], color_normalized[2], 1.0
-                elif len(color_normalized) == 4:
-                    clear_color = color_normalized
-                else:
-                    raise ValueError("Color should be a 3 or 4 component tuple")
-
-            gl.glClearColor(*clear_color)
-
-            if self.depth_attachment:
-                if self._ctx.gl_api == "gl":
-                    gl.glClearDepth(depth)
-                else:  # gles only supports glClearDepthf
-                    gl.glClearDepthf(depth)
-
-                gl.glClear(gl.GL_COLOR_BUFFER_BIT | gl.GL_DEPTH_BUFFER_BIT)
-            else:
-                gl.glClear(gl.GL_COLOR_BUFFER_BIT)
-
-            self.scissor = scissor_values
-
+    @abstractmethod
     def read(self, *, viewport=None, components=3, attachment=0, dtype="f1") -> bytes:
         """
         Read the raw framebuffer pixels.
@@ -409,35 +287,7 @@ class Framebuffer:
             dtype:
                 The data type to read. Pixel data will be converted to this format.
         """
-        # TODO: use texture attachment info to determine read format?
-        try:
-            frmt = pixel_formats[dtype]
-            base_format = frmt[0][components]
-            pixel_type = frmt[2]
-            component_size = frmt[3]
-        except Exception:
-            raise ValueError(f"Invalid dtype '{dtype}'")
-
-        with self.activate():
-            # Configure attachment to read from. Does not work on default framebuffer.
-            if not self.is_default:
-                gl.glReadBuffer(gl.GL_COLOR_ATTACHMENT0 + attachment)
-
-            gl.glPixelStorei(gl.GL_PACK_ALIGNMENT, 1)
-            gl.glPixelStorei(gl.GL_UNPACK_ALIGNMENT, 1)
-
-            if viewport:
-                x, y, width, height = viewport
-            else:
-                x, y, width, height = 0, 0, *self.size
-
-            data = (gl.GLubyte * (components * component_size * width * height))(0)
-            gl.glReadPixels(x, y, width, height, base_format, pixel_type, data)
-
-            if not self.is_default:
-                gl.glReadBuffer(gl.GL_COLOR_ATTACHMENT0)  # Reset to default
-
-        return string_at(data, len(data))
+        raise NotImplementedError("The enabled graphics backend does not support this method.")
 
     def resize(self):
         """
@@ -448,31 +298,14 @@ class Framebuffer:
         self._width, self._height = self._detect_size()
         self.viewport = 0, 0, self.width, self._height
 
+    @abstractmethod
     def delete(self):
         """
         Destroy the underlying OpenGL resource.
 
         .. warning:: Don't use this unless you know exactly what you are doing.
         """
-        Framebuffer.delete_glo(self._ctx, self._glo)
-        self._glo.value = 0
-
-    @staticmethod
-    def delete_glo(ctx, framebuffer_id):
-        """
-        Destroys the framebuffer object
-
-        Args:
-            ctx:
-                The context this framebuffer belongs to
-            framebuffer_id:
-                Framebuffer id destroy (glo)
-        """
-        if gl.current_context is None:
-            return
-
-        gl.glDeleteFramebuffers(1, framebuffer_id)
-        ctx.stats.decr("framebuffer")
+        raise NotImplementedError("The enabled graphics backend does not support this method.")
 
     def _detect_size(self) -> tuple[int, int]:
         """Detect the size of the framebuffer based on the attachments"""
@@ -492,36 +325,7 @@ class Framebuffer:
                 raise ValueError("All framebuffer attachments should have the same size")
         return expected_size
 
-    @staticmethod
-    def _check_completeness() -> None:
-        """
-        Checks the completeness of the framebuffer.
-
-        If the framebuffer is not complete, we cannot continue.
-        """
-        # See completeness rules : https://www.khronos.org/opengl/wiki/Framebuffer_Object
-        states = {
-            gl.GL_FRAMEBUFFER_UNSUPPORTED: "Framebuffer unsupported. Try another format.",
-            gl.GL_FRAMEBUFFER_INCOMPLETE_ATTACHMENT: "Framebuffer incomplete attachment.",
-            gl.GL_FRAMEBUFFER_INCOMPLETE_MISSING_ATTACHMENT: "Framebuffer missing attachment.",
-            gl.GL_FRAMEBUFFER_INCOMPLETE_DIMENSIONS_EXT: "Framebuffer unsupported dimension.",
-            gl.GL_FRAMEBUFFER_INCOMPLETE_FORMATS_EXT: "Framebuffer incomplete formats.",
-            gl.GL_FRAMEBUFFER_INCOMPLETE_DRAW_BUFFER: "Framebuffer incomplete draw buffer.",
-            gl.GL_FRAMEBUFFER_INCOMPLETE_READ_BUFFER: "Framebuffer incomplete read buffer.",
-            gl.GL_FRAMEBUFFER_COMPLETE: "Framebuffer is complete.",
-        }
-
-        status = gl.glCheckFramebufferStatus(gl.GL_FRAMEBUFFER)
-        if status != gl.GL_FRAMEBUFFER_COMPLETE:
-            raise ValueError(
-                "Framebuffer is incomplete. {}".format(states.get(status, "Unknown error"))
-            )
-
-    def __repr__(self):
-        return "<Framebuffer glo={}>".format(self._glo.value)
-
-
-class DefaultFrameBuffer(Framebuffer):
+class DefaultFrameBuffer(Framebuffer, ABC):
     """
     Represents the default framebuffer.
 
@@ -539,8 +343,6 @@ class DefaultFrameBuffer(Framebuffer):
     is_default = True
     """Is this the default framebuffer? (window buffer)"""
 
-    __slots__ = ()
-
     def __init__(self, ctx: "Context"):
         self._ctx = ctx
         # TODO: Can we query this?
@@ -550,10 +352,6 @@ class DefaultFrameBuffer(Framebuffer):
         self._depth_attachment = None
         self._depth_mask = True
 
-        value = c_int()
-        gl.glGetIntegerv(gl.GL_DRAW_FRAMEBUFFER_BINDING, value)
-        self._glo = gl.GLuint(value.value)
-
         # Query draw buffers. This will most likely return GL_BACK
         # value = c_int()
         # gl.glGetIntegerv(gl.GL_DRAW_BUFFER0, value)
@@ -561,16 +359,6 @@ class DefaultFrameBuffer(Framebuffer):
         # self._draw_buffers = (gl.GLuint * 1)(value.value)
         # NOTE: Don't query for now
         self._draw_buffers = None
-
-        # Query viewport values by inspecting the scissor box
-        values = (c_int * 4)()
-        gl.glGetIntegerv(gl.GL_SCISSOR_BOX, values)
-        x, y, width, height = list(values)
-
-        self._viewport = x, y, width, height
-        self._scissor = None
-        self._width = width
-        self._height = height
 
         # HACK: Signal the default framebuffer having depth buffer
         self._depth_attachment = True  # type: ignore
@@ -594,7 +382,8 @@ class DefaultFrameBuffer(Framebuffer):
         """Get the framebuffer size of the window"""
         return self._ctx.window.get_framebuffer_size()
 
-    def _get_viewport(self) -> tuple[int, int, int, int]:
+    @Framebuffer.viewport.getter
+    def viewport(self) -> tuple[int, int, int, int]:
         """
         Get or set the framebuffer's viewport.
         The viewport parameter are ``(x, y, width, height)``.
@@ -617,31 +406,13 @@ class DefaultFrameBuffer(Framebuffer):
             int(self._viewport[3] / ratio),
         )
 
-    def _set_viewport(self, value: tuple[int, int, int, int]):
-        if not isinstance(value, tuple) or len(value) != 4:
-            raise ValueError("viewport should be a 4-component tuple")
+    @Framebuffer.viewport.setter
+    @abstractmethod
+    def viewport(self, value: tuple[int, int, int, int]):
+        raise NotImplementedError("The enabled graphics backend does not support this method.")
 
-        ratio = self.ctx.window.get_pixel_ratio()
-        self._viewport = (
-            int(value[0] * ratio),
-            int(value[1] * ratio),
-            int(value[2] * ratio),
-            int(value[3] * ratio),
-        )
-
-        # If the framebuffer is bound we need to set the viewport.
-        # Otherwise it will be set on use()
-        if self._ctx.active_framebuffer == self:
-            gl.glViewport(*self._viewport)
-            if self._scissor is None:
-                # FIXME: Probably should be set to the framebuffer size
-                gl.glScissor(*self._viewport)
-            else:
-                gl.glScissor(*self._scissor)
-
-    viewport = property(_get_viewport, _set_viewport)
-
-    def _get_scissor(self) -> tuple[int, int, int, int] | None:
+    @Framebuffer.scissor.getter
+    def scissor(self) -> tuple[int, int, int, int] | None:
         """
         Get or set the scissor box for this framebuffer.
 
@@ -667,24 +438,7 @@ class DefaultFrameBuffer(Framebuffer):
             int(self._scissor[3] / ratio),
         )
 
-    def _set_scissor(self, value):
-        if value is None:
-            # FIXME: Do we need to reset something here?
-            self._scissor = None
-            if self._ctx.active_framebuffer == self:
-                gl.glScissor(*self._viewport)
-        else:
-            ratio = self.ctx.window.get_pixel_ratio()
-            self._scissor = (
-                int(value[0] * ratio),
-                int(value[1] * ratio),
-                int(value[2] * ratio),
-                int(value[3] * ratio),
-            )
-
-            # If the framebuffer is bound we need to set the scissor box.
-            # Otherwise it will be set on use()
-            if self._ctx.active_framebuffer == self:
-                gl.glScissor(*self._scissor)
-
-    scissor = property(_get_scissor, _set_scissor)
+    @Framebuffer.scissor.setter
+    @abstractmethod
+    def scissor(self, value):
+        raise NotImplementedError("The enabled graphics backend does not support this method.")
